requirements.txt
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
decorator==4.4.0
importlib-metadata==0.19
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
more-itertools==7.2.0
packaging==19.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.12.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.2
pytest==5.0.1
python-dateutil==2.8.0
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
zipp==0.5.2
